refactor(config): move loadConfig function to config module; streamline client configuration loading
