+++ src/salt.ts
@@ -1,27 +1,10 @@
-import fs from "fs";
-import path from "path";
-import yaml from "yaml";
 import discord from "discord.js";
 
 import Logger from "./utils/logger";
 import { Command, SlashCommand } from "./types";
 import CommandLogger from "./core/command/logger";
+import { loadConfig } from "./utils/config";
 
-
-/**
- * Loads configuration from YAML file
- * @returns Configuration object
- */
-const loadConfig = (client: discord.Client) => {
-    try {
-        const configPath = path.join(__dirname, "../config/config.yml");
-        const file = fs.readFileSync(configPath, "utf8");
-        return yaml.parse(file);
-    } catch (error) {
-        client.logger.error(`[SALT] Failed to load configuration: ${error}`);
-        process.exit(1);
-    }
-};
 
 /**
  * Creates and configures the Discord client with all necessary properties
